Move some generically useful view code into the experiments application
